#4 Refactor: Timestamped 모델 재사용성있게 분리
