+++ wtnt/user/models.py
@@ -1,16 +1,10 @@
 from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 from django.db import models
 from .managers import UserManager
+from core.models import TimestampedModel
 
 
 # Create your models here.
-class TimestampedModel(models.Model):
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-    class Meta:
-        abstract = True
-        ordering = ["-created_at", "-updated_at"]
 
 
 class CustomUser(AbstractBaseUser, PermissionsMixin, TimestampedModel):
